First working (almost) version.
After calculation of hash for each block, we iterate through
them to calculate further hashes, till are left with only one
hash which is the root hash.

This is not a perfect implementation of Merkle tree because
the approach we are using will not result in a balanced binary
tree.